Add bookmark and watched percent support to videos
Add support for displaying a bookmark state within
the video manager. This uses the playback state class
for caching the playback state for all videos.
Also adds support for displaying watched progress
state to video manager. This needs the last playing
position (already supported) as well as a total
frames markup datum within the filemarkup table.
The watched progress is then calculated and can be
displayed via a progress in the currently selected list
item.