Fix displaying watched state after playback
When watching a video to the end, the player can mark
(if enabled by a setting) the video as watched. This
was not reflected in the GUI state until now.
So after receiving the playback stopped event, we load
the video metadata of the current video and sync it with
the metadata from the cache.